+++ retz-common/src/main/java/io/github/retz/protocol/Request.java
@@ -40,6 +40,7 @@ public abstract class Request {
     public static final String PUT = "PUT";
     public static final String GET = "GET";
     public static final String DELETE = "DELETE";
+    public static final String POST = "POST";
     public abstract String resource();
     public abstract String method();
     public abstract boolean hasPayload();