+++ retz-common/src/main/java/io/github/retz/protocol/ScheduleRequest.java
@@ -43,7 +43,7 @@ public class ScheduleRequest extends Request {
 
     @Override
     public String method() {
-        return PUT;
+        return POST;
     }
 
     @Override